addons: Move source files to internal directory
