+++ src/addons/internal/AddonItemModel.cpp
