+++ src/addons/internal/AddonItemModel.h
