+++ src/addons/internal/AddonListModel.cpp
