+++ src/addons/internal/AddonModel.cpp
