+++ src/addons/internal/AddonModel.h
